Rename package to material3
