+++ demos/appyx-navigation/common/src/commonMain/kotlin/com/bumble/appyx/navigation/node/container/MainNavNode.kt
@@ -2,7 +2,7 @@ package com.bumble.appyx.navigation.node.container
 
 import com.bumble.appyx.navigation.MainNavItem
 import com.bumble.appyx.navigation.modality.BuildContext
-import com.bumble.appyx.utils.material.AppyxMaterial3NavNode
+import com.bumble.appyx.utils.material3.AppyxMaterial3NavNode
 
 
 class MainNavNode(
